WebWorker is now Worker
